consulta-natural-service/requirements.txt
fastapi
uvicorn
python-dotenv
openai==0.28
firebase-admin
google-generativeai